Prevent reloads while picking/uploading images
Some combination of the picker showing and the image resizing is causing memory pressure, which would result in a trim event being fired and the page being reloaded in the middle of the upload process.

This does a few things to address that:

- Update Crosswalk to the latest stable (22.52.561.4)
- Set the `android:largeheap` setting in the manifest
- Use the `RGB_565` setting for image reprocessing in Picasso
- Prevent `trimMemory` events from triggering a reload while picking images

It does now seem that the Crosswalk picker does not return the image capture intent back properly in a way that the image gets processed, which appears to be unrelated but could also cause memory pressure.